deps(python): Fix Security Vulnerabilities (#263)
# Pull Request

## Description

This pull request includes updates to dependencies in the `Dockerfile`
and `pyproject.toml` files to ensure the use of the latest versions. The
most important changes include updating the git version in the
Dockerfile and modifying several dependencies in the `pyproject.toml`
file.

Dependency updates:

*
[`Dockerfile`](diffhunk://#diff-dd2c0eb6ea5cfc6c4bd4eac30934e2d5746747af48fef6da689e85b752f39557L16-R16):
Updated the git version from `2.45.2-r0` to `2.47.1-r0`.
*
[`pyproject.toml`](diffhunk://#diff-50c86b7ed8ac2cf95bd48334961bf0530cdc77b5a56f852c5c61b89d735fd711L17-R20):
Updated the `ruff` version from `0.8.0` to `0.8.2`.
*
[`pyproject.toml`](diffhunk://#diff-50c86b7ed8ac2cf95bd48334961bf0530cdc77b5a56f852c5c61b89d735fd711L17-R20):
Updated the `check-jsonschema` version from `0.29.4` to `0.30.0`.