ci: Fix Local Action Timeouts (#267)
# Pull Request

## Description

This pull request includes a small but important change to the
`.github/workflows/code-checks.yml` file. The change involves adding a
GitHub token to the `Run GitHub Stats Analyser` job to ensure it has the
necessary permissions to execute.

*
[`.github/workflows/code-checks.yml`](diffhunk://#diff-ddf88e15b08104435ae66be9982938335f6c290a85de4cb9a09868e0e01dd4d4R239):
Added `GITHUB_TOKEN` to the `Run GitHub Stats Analyser` job to provide
necessary permissions.