feat!: Reformat JSON Output and Add Repository Owner (#265)
# Pull Request

## Description

This pull request includes significant changes to the
`analyser/statistics.py` file, the corresponding tests, and the schema
validation for repository statistics. The main changes involve the
introduction of a new function for generating output files and updating
the tests and schema to accommodate this new functionality.

### Key Changes:

#### New Functionality:
* Added a new function `generate_output_file` in
`analyser/statistics.py` to handle the creation of output files. This
function takes a configuration and a dataframe as arguments and writes
the data to a JSON file.

#### Refactoring:
* Refactored the `create_statistics` function to use the new
`generate_output_file` function instead of directly writing to a JSON
file.

#### Testing:
* Updated the test file `analyser/tests/test_statistics.py` to include
tests for the new `generate_output_file` function. This includes adding
a patch for `generate_output_file` in existing tests and creating a new
test specifically for this function.
[[1]](diffhunk://#diff-96ff6623a4ce15f9e281c09ed90d0b0c46f7a01c7cff18ad2e0a821b22fb5e15L3-R12)
[[2]](diffhunk://#diff-96ff6623a4ce15f9e281c09ed90d0b0c46f7a01c7cff18ad2e0a821b22fb5e15R24)
[[3]](diffhunk://#diff-96ff6623a4ce15f9e281c09ed90d0b0c46f7a01c7cff18ad2e0a821b22fb5e15R53)
[[4]](diffhunk://#diff-96ff6623a4ce15f9e281c09ed90d0b0c46f7a01c7cff18ad2e0a821b22fb5e15R69-R88)

#### Schema Validation:
* Modified the `repository_statistics_schema.json` file to include new
properties for `repository_owner` and `repositories`, ensuring that the
JSON output adheres to the updated schema.
[[1]](diffhunk://#diff-f2ae76f93583821645527d189948b6dbb8e68c831573dd54d4b3483d172de4e8R4-R9)
[[2]](diffhunk://#diff-f2ae76f93583821645527d189948b6dbb8e68c831573dd54d4b3483d172de4e8R41-R42)

These changes enhance the modularity and test coverage of the codebase,
making it more maintainable and robust.

Fixes #264