build: Only Write GitHub Action Summary when Action is run in GitHub Actions (#287)
# Pull Request

## Description

This pull request includes changes to the `generate_action_summary`
function and adds tests to ensure its proper functionality. The primary
goal is to improve the handling of the action summary generation based
on the environment.

Enhancements to action summary generation:

*
[`analyser/utils/action_summary.py`](diffhunk://#diff-4a11a4c00a6233aa1433fc099851e5a1840eb6291b142e918a58ae20bbc2566cR16-R22):
Modified the `generate_action_summary` function to check if it is
running in GitHub Actions and generate the action summary accordingly.
This includes logging messages for both scenarios.

Addition of tests for action summary:

*
[`analyser/utils/tests/test_action_summary.py`](diffhunk://#diff-2256e8f8f5aeb6f380e350743483f68a9d626b90f1307bb89ba44f582d6538beR1-R37):
Added tests to verify the behavior of `generate_action_summary` when
running in GitHub Actions and when not running in GitHub Actions. This
includes using mocks to simulate the environment and file operations.

Fixes #257