chore: Add Code of Conduct (#288)
# Pull Request

## Description


This pull request includes the addition of a new Code of Conduct
document to the project. The document outlines the expected behavior and
responsibilities of community members, as well as the enforcement
guidelines for handling violations.

Key changes:

* Added `docs/CODE_OF_CONDUCT.md` file with the Contributor Covenant
Code of Conduct, including sections on our pledge, standards,
enforcement responsibilities, scope, enforcement process, and guidelines
for community impact.

Fixes #289