ci: Scan GitHub Actions with CodeQL (#303)
# Pull Request

## Description

This pull request includes changes to the
`.github/workflows/code-checks.yml` file to enhance the flexibility and
efficiency of the CI/CD pipeline. The most important change is the
introduction of a strategy matrix to support multiple languages in the
CodeQL analysis.

Enhancements to CI/CD pipeline:

*
[`.github/workflows/code-checks.yml`](diffhunk://#diff-ddf88e15b08104435ae66be9982938335f6c290a85de4cb9a09868e0e01dd4d4R86-R98):
Added a strategy matrix to support both Python and GitHub Actions
languages in the CodeQL analysis.
*
[`.github/workflows/code-checks.yml`](diffhunk://#diff-ddf88e15b08104435ae66be9982938335f6c290a85de4cb9a09868e0e01dd4d4R86-R98):
Updated the `Checkout repository` step name to `Checkout` for brevity.
*
[`.github/workflows/code-checks.yml`](diffhunk://#diff-ddf88e15b08104435ae66be9982938335f6c290a85de4cb9a09868e0e01dd4d4R86-R98):
Modified the `languages` parameter in the `Initialize CodeQL` step to
use the matrix variable.

Fixes #302