chore: Set markdown-link-checker job runner to ubuntu-22.04 (#306)
# Pull Request

## Description

This pull request includes updates to the GitHub Actions workflow and
the Dockerfile to ensure compatibility with newer versions of the
operating system and software packages. The most important changes
include updating the runner version for GitHub Actions and updating the
Git version in the Dockerfile.

Updates to GitHub Actions workflow:

*
[`.github/workflows/code-checks.yml`](diffhunk://#diff-ddf88e15b08104435ae66be9982938335f6c290a85de4cb9a09868e0e01dd4d4L105-R105):
Changed the runner version from `ubuntu-latest` to `ubuntu-22.04` for
the `check-markdown-links` job.
*
[`.github/workflows/code-checks.yml`](diffhunk://#diff-ddf88e15b08104435ae66be9982938335f6c290a85de4cb9a09868e0e01dd4d4L214-R214):
Changed the runner version from `ubuntu-latest` to `ubuntu-22.04` for
the `run-local-action` job.

Updates to Dockerfile:

*
[`Dockerfile`](diffhunk://#diff-dd2c0eb6ea5cfc6c4bd4eac30934e2d5746747af48fef6da689e85b752f39557L18-R18):
Updated the Git version from `2.47.1-r0` to `2.47.2-r0` to ensure
compatibility with the latest features and security updates.