ci: Upgrade Markdown Link Checker (#309)
# Pull Request

## Description

This pull request includes updates to the GitHub Actions workflow for
checking Markdown links. The changes ensure that the workflow uses the
latest versions of the runner and the action.

Updates to GitHub Actions workflow:

*
[`.github/workflows/code-checks.yml`](diffhunk://#diff-ddf88e15b08104435ae66be9982938335f6c290a85de4cb9a09868e0e01dd4d4L107-R115):
Updated the runner to `ubuntu-latest` instead of a specific version
(`ubuntu-22.04`).
*
[`.github/workflows/code-checks.yml`](diffhunk://#diff-ddf88e15b08104435ae66be9982938335f6c290a85de4cb9a09868e0e01dd4d4L107-R115):
Updated the `UmbrellaDocs/action-linkspector` action to version `v1.2.5`
from `v1.2.4`.