ci: Regularly Clear Caches (#323)
# Pull Request

## Description

This pull request introduces a new GitHub Actions workflow to clean
caches periodically. The most important change is the addition of the
`clean-caches.yml` file, which schedules a job to delete GitHub Action
caches on a monthly basis.

New GitHub Actions workflow:

*
[`.github/workflows/clean-caches.yml`](diffhunk://#diff-d0394e4336a74cdfc1d4cff05d056b893ac7ff922eacf4448e104a754f386b8dR1-R24):
Added a new workflow named "Clean Caches" that runs on a schedule (cron
job) and can be manually triggered. It includes steps to check out the
repository and delete all caches using the GitHub CLI.