chore: Add Contributing File (#334)
# Pull Request

## Description

This pull request includes a couple of important changes to the codebase
and documentation. The changes improve the readability of the code and
provide guidelines for contributing to the project.

Code readability improvement:

*
[`analyser/file_analysis/programming_language_analysis.py`](diffhunk://#diff-f1f475fcddb60c6815991e17b0ebcddbbf4485c8c72890fd527aee818d6e50b2R101-L103):
Removed an unnecessary logging statement within the `count_sloc`
function to streamline the code.

Documentation update:

*
[`docs/CONTRIBUTING.md`](diffhunk://#diff-4c6b93aa75d5affde60dc3849606c9acd75ed444d52e99f3055fc0c7aa77e9e0R1-R3):
Added a new section to encourage contributions by providing guidelines
for raising issues and submitting pull requests.