chore: Add Pinact Update Command (#383)
# Pull Request

## Description

This pull request includes updates to dependencies and configuration
files to ensure compatibility and add new functionality. The most
important changes include resolving a conflict in the Markdown
link-checking workflow, adding a new `pinact-update` command, and
updating the minimum version requirement for `lefthook`.

### Dependency and Workflow Updates:

*
[`.github/workflows/code-checks.yml`](diffhunk://#diff-ddf88e15b08104435ae66be9982938335f6c290a85de4cb9a09868e0e01dd4d4R132-R136):
Resolved a conflict in the `Check Markdown links` workflow by updating
the `UmbrellaDocs/action-linkspector` action to version `v1.3.4`.

*
[`lefthook.yml`](diffhunk://#diff-ad6a01e589b8b1b214ca310dbb8d2e4314f6c612b921050c73c97455de43884dL2-R2):
Updated the `min_version` of `lefthook` from `1.11.4` to `1.11.11` to
ensure compatibility with the latest features and fixes.

### New Command Addition:

*
[`Justfile`](diffhunk://#diff-2f90408c2b0302b1cdb7f5d634114750837f940fa82d13057d9c18d0170a7e5cR182-R185):
Added a new `pinact-update` command to allow running `pinact` with the
`--update` option for updating configurations.