chore: Remove Pinact Configuration (#403)
# Pull Request

## Description

This pull request removes the use of a custom configuration file for
`pinact` and updates the related commands in the `Justfile` to use the
default configuration. The most important changes are as follows:

### Removal of custom configuration:

* Deleted the `.github/other-configurations/pinact.yml` file, which
contained a custom configuration for `pinact`. This includes removing
the schema reference, version, and ignored actions.

### Updates to `Justfile` commands:

* Modified `pinact-run`, `pinact-check`, and `pinact-update` commands in
the `Justfile` to no longer specify the custom configuration file and
instead rely on the default configuration.