chore: Set Python Version (#436)
# Pull Request

## Description

This pull request includes updates to dependencies and configuration
files to ensure compatibility and improve stability. The most important
changes include updating the `uv` library version, modifying the Python
version requirement, and updating the minimum version for the `lefthook`
tool.

### Dependency Updates:
* Updated the `uv` library version from `0.7.3` to `0.7.5` in both the
`Dockerfile` and `pyproject.toml` to ensure compatibility with the
latest features and fixes.
[[1]](diffhunk://#diff-dd2c0eb6ea5cfc6c4bd4eac30934e2d5746747af48fef6da689e85b752f39557L8-R8)
[[2]](diffhunk://#diff-50c86b7ed8ac2cf95bd48334961bf0530cdc77b5a56f852c5c61b89d735fd711L30-R30)

### Configuration Updates:
* Changed the Python version requirement in `pyproject.toml` from
`>=3.13` to `~=3.13` to better align with semantic versioning practices.
* Updated the minimum required version of `lefthook` from `1.11.12` to
`1.11.13` in `lefthook.yml`.