deps(python): Bump UV Version (#442)
# Pull Request

## Description

This pull request updates the version of the `uv` dependency across the
project to ensure compatibility and access to the latest features. The
changes include modifying the `Dockerfile` and `pyproject.toml` to
reflect the new version.

Dependency version update:

*
[`Dockerfile`](diffhunk://#diff-dd2c0eb6ea5cfc6c4bd4eac30934e2d5746747af48fef6da689e85b752f39557L8-R8):
Updated the `uv` dependency from version `0.7.5` to `0.7.11` in the
`builder` stage.
*
[`pyproject.toml`](diffhunk://#diff-50c86b7ed8ac2cf95bd48334961bf0530cdc77b5a56f852c5c61b89d735fd711L30-R30):
Changed the `required-version` of `uv` from `0.7.5` to `0.7.11` in the
`[tool.uv]` section.