ci: Update Pull Request Tasks Permissions (#562)
# Pull Request

## Description

This pull request makes a minor update to the permissions structure in
the `.github/workflows/pull-request-tasks.yml` file. The main change is
to move the `pull-requests: read` permission from the global workflow
level to the specific job level for `check-pull-request-title`.

Workflow permissions update:

* The global `permissions` block is replaced with an empty object, and
the `pull-requests: read` permission is now set only for the
`check-pull-request-title` job. This scopes down the permission to just
where it's needed, following GitHub Actions best practices.